refactor: :recycle: move metadata checks to metadata class
