+++ src/zen_do/examples.py
@@ -8,7 +8,9 @@ from zen_do.zenodo_metadata import (
 )
 
 
-def example_metadata(title: str = "Test Book") -> ZenodoMetadata:
+def example_metadata(
+    title: str = "Test Book", urn: str = "urn:zenodo:my-org:project:book"
+) -> ZenodoMetadata:
     """A set of example Zenodo metadata."""
     return ZenodoMetadata(
         title=title,
@@ -20,7 +22,7 @@ def example_metadata(title: str = "Test Book") -> ZenodoMetadata:
         ],
         related_identifiers=[
             ZenodoRelatedIdentifier(
-                identifier="urn:zenodo:my-org:project:book",
+                identifier=urn,
                 relation="isIdenticalTo",
                 resource_type="other",
                 scheme="urn",
