+++ tests/test_zenodo_get_deposit.py
@@ -1,7 +1,6 @@
-import json
 from pathlib import Path
 
-from pytest import MonkeyPatch, fixture, mark, raises
+from pytest import MonkeyPatch, fixture, raises
 
 from zen_do.examples import example_deposit, example_metadata
 from zen_do.zenodo_get_deposit import zenodo_get_deposit
@@ -77,54 +76,3 @@ def test_returns_none_if_no_matching_deposits(_zenodo_json):
     deposit = zenodo_get_deposit(deposits)
 
     assert deposit is None
-
-
-def test_raises_error_if_zenodo_json_has_no_urn_id(monkeypatch, tmp_path):
-    metadata = example_metadata()
-    monkeypatch.chdir(tmp_path)
-    del metadata.related_identifiers[0]
-    (tmp_path / ".zenodo.json").write_text(metadata.model_dump_json())
-
-    with raises(ValueError):
-        zenodo_get_deposit([])
-
-
-def test_raises_error_if_zenodo_json_has_multiple_urn_ids(monkeypatch, tmp_path):
-    metadata = example_metadata()
-    monkeypatch.chdir(tmp_path)
-    metadata.related_identifiers.append(
-        ZenodoRelatedIdentifier(
-            identifier="urn:zenodo:my-org:project:poster",
-            relation="isIdenticalTo",
-            resource_type="other",
-            scheme="urn",
-        )
-    )
-    (tmp_path / ".zenodo.json").write_text(metadata.model_dump_json())
-
-    with raises(ValueError):
-        zenodo_get_deposit([])
-
-
-@mark.parametrize(
-    "urn",
-    [
-        "",
-        "not a URN",
-        "urn",
-        "urn:",
-        "urn:unknown",
-        "urn:zenodo",
-        "urn:zenodo:",
-        "urn:zenodo:a:",
-        "urn:zenodo:a/b",
-    ],
-)
-def test_flags_incorrect_urn(monkeypatch, tmp_path, urn):
-    metadata_json = example_metadata().model_dump()
-    metadata_json["related_identifiers"][0]["identifier"] = urn
-    monkeypatch.chdir(tmp_path)
-    (tmp_path / ".zenodo.json").write_text(json.dumps(metadata_json))
-
-    with raises(ValueError):
-        zenodo_get_deposit([])
